Fix bug with adding content models
This was passing an extra layer of array nesting to
`islandora_solr_metadata_add_content_models` in the `$content_models`
array.

This caused issues when using postgres, because its
database engine uses the `bindParams` method to insert values, which
complains about an array to string conversion. The end result
is that the literal string "Array" gets inserted into the database table
instead of the cmodel associations.